grsec: Apply grsecurity-2.9.1-3.2.50-201309011629.patch
commit 2247ff86d4ee11fb96e70bea4b2c60a52e03974f
Merge: 6414043 0c7e99e
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sun Sep 1 15:15:57 2013 -0400

    Merge branch 'pax-stable2' into grsec-stable2

commit 0c7e99e26011b88ded65c1c1ebeb715193620827
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sun Sep 1 15:15:13 2013 -0400

    Update to pax-linux-3.2.50-test118.patch:
    - fixed a REFCOUNT false positive, by Mathias Krause <minipli@googlemail.com>
    - fixed a bunch more after a quick audit of atomic_inc_return users
    - fixed a few more REFCOUNT false positives, by Mathias Krause <minipli@googlemail.com>
    - got inet_getid and ipv6_select_ident rid of the cmpxchg loop

 drivers/crypto/hifn_795x.c             |    4 ++--
 drivers/edac/edac_device.c             |    4 ++--
 drivers/edac/edac_pci.c                |    4 ++--
 drivers/firewire/core-card.c           |    4 ++--
 drivers/input/serio/serio_raw.c        |    4 ++--
 drivers/media/rc/rc-main.c             |    4 ++--
 drivers/media/video/ivtv/ivtv-driver.c |    2 +-
 drivers/media/video/v4l2-device.c      |    4 ++--
 drivers/net/usb/sierra_net.c           |    4 ++--
 drivers/regulator/core.c               |    4 ++--
 drivers/tty/hvc/hvsi.c                 |   14 +++++++-------
 drivers/tty/hvc/hvsi_lib.c             |    6 +++---
 drivers/tty/serial/ioc4_serial.c       |    6 +++---
 drivers/tty/serial/msm_serial.c        |    4 ++--
 drivers/usb/misc/appledisplay.c        |    4 ++--
 fs/afs/inode.c                         |    4 ++--
 fs/fscache/cookie.c                    |    4 ++--
 include/media/v4l2-device.h            |    2 +-
 include/net/inetpeer.h                 |   13 +++++--------
 kernel/trace/trace_clock.c             |    4 ++--
 net/ipv6/ip6_output.c                  |   15 ++++++---------
 net/xfrm/xfrm_state.c                  |    4 ++--
 security/selinux/avc.c                 |    6 +++---
 23 files changed, 59 insertions(+), 65 deletions(-)

commit 6414043ca36390f651f66a33eb5e1c7a553efa0b
Merge: ef56283 70a6353
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sat Aug 31 21:05:24 2013 -0400

    Merge branch 'pax-stable2' into grsec-stable2

commit 70a6353af67a7a54baf9c659bb4486b3fd1b1143
Author: Brad Spengler <spender@grsecurity.net>
Date:   Sat Aug 31 21:04:42 2013 -0400

    Update to pax-linux-3.2.50-test117.patch:
    - removed unnecessary mark_sym_for_renaming calls from the gcc plugins, reported by Emese Revfy
    - __copy_from_user_inatomic on amd64 will now return unsigned long like other userland accessors do
    - inspired by Dan Carpenter's recent fix (http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=909bd5926d474e275599094acad986af79671ac9)
      Emese Revfy wrote a gcc plugin to find other instances of the same error, here's the fallout
      (come to the 10th H2HC if you want to learn about the magic behind this and other plugins):
      - icmpv6_filter: no memory corruption, probably just some logical error in the caller
      - dccp_new/dccp_packet/dccp_error: probably remote kernel stack overflow (12 byte network data overwriting a local ptr variable)
      - gigaset_brkchars: causes DMA on the kernel stack, some archs don't like it (more of this is to come)
      - isdn_ioctl/IIOCDBGVAR: kernel heap address leak (by design), restricted to CAP_SYS_RAWIO now
      - lowpan_header_create: leaks 3 bytes of a kernel heap address over the network
    - fixed a few more format strings
    - audited and fixed arm and sparc for proper atomic_unchecked_t usage

 arch/sparc/kernel/smp_64.c                 |   12 ++++++------
 arch/sparc/kernel/traps_64.c               |   14 +++++++-------
 arch/sparc/mm/init_64.c                    |   10 +++++-----
 arch/x86/include/asm/uaccess_64.h          |    2 +-
 drivers/isdn/gigaset/usb-gigaset.c         |    2 +-
 drivers/isdn/i4l/isdn_common.c             |    2 ++
 drivers/net/wireless/hostap/hostap_ioctl.c |    4 ++--
 drivers/platform/x86/wmi.c                 |    2 +-
 drivers/scsi/sd.c                          |    2 +-
 fs/ntfs/file.c                             |    2 +-
 fs/ntfs/super.c                            |    6 +++---
 kernel/events/internal.h                   |    2 +-
 kernel/futex.c                             |    2 +-
 mm/filemap.c                               |    8 ++++----
 net/ieee802154/6lowpan.c                   |    2 +-
 net/ipv6/raw.c                             |    2 +-
 net/netfilter/nf_conntrack_proto_dccp.c    |    6 +++---
 sound/pci/hda/hda_codec.c                  |    8 ++------
 tools/gcc/kernexec_plugin.c                |   18 ++++++++++++------
 tools/gcc/latent_entropy_plugin.c          |   26 ++++++++++----------------
 tools/gcc/size_overflow_plugin.c           |    3 +--
 21 files changed, 66 insertions(+), 69 deletions(-)

Signature-tree: 779ed53645862e9672ae296fba18926c6615f42e

+++ kernel/events/internal.h
@@ -78,7 +78,7 @@ static unsigned long perf_data_size(struct ring_buffer *rb)
 
 static inline void
 __output_copy(struct perf_output_handle *handle,
-		   const void *buf, unsigned int len)
+		   const void *buf, unsigned long len)
 {
 	do {
 		unsigned long size = min_t(unsigned long, handle->size, len);
